refactor(cli): split index.ts entry point (Phase 1 #4 batch 4b)
1,188 → 66 LOC entry point + 9 sub-files under src/cli/.

Splits the CLI entry into a thin dispatcher that delegates to focused
modules: agentShorthand (alias resolution), helpScreen (usage output),
version, commandRouter, and per-command handlers (pair, start, status,
costs, simple). 33 new tests covering agent alias normalization and the
help-screen rendering path.

Every command flow — pair, start, status, costs, and the pass-through
simple agents — is exercised by the existing 1,346 CLI unit tests.
